Refactoring (separating) how the symfony config is brought into the parser class.
The new change makes the config and the parser class separate things. A new static method and accessor on the plugin configuration is the correct way to access the filter and is responsible for merging in the configuration where necessary.

I'm on a recent kick of seperating the symfony config from classes - they should be passed in as dependencies. Otherwise, your classes are hopelessly dependent on symfony's static configuration - in other words, a global dependency.